Move bow parser to args
